--- ordering/src/main/java/com/careup/ordering/common/BaseTimeEntity.java
@@ -1,19 +1,0 @@
-package com.careup.ordering.common;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@Getter
-public abstract class BaseTimeEntity {
-
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-}
