--- ordering/src/main/java/com/careup/ordering/member/entity/Gender.java
@@ -1,6 +1,0 @@
-package com.careup.ordering.member.entity;
-
-
-public enum Gender {
-    M, W
-}
