--- ordering/src/main/java/com/careup/ordering/order/entity/OrderType.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.order.entity;
-
-public enum OrderType {
-    ONLINE, OFFLINE
-}
