+++ ordering/src/main/java/com/careup/ordering/ordering/order/entity/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.careup.ordering.order.entity;
+package com.careup.ordering.ordering.order.entity;
 
 public enum OrderStatus {
     PENDING, CONFIRMED, CANCELLED
