+++ ordering/src/main/java/com/careup/ordering/ordering/order/entity/OrderType.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.order.entity;
+
+public enum OrderType {
+    ONLINE, OFFLINE
+}
