+++ ordering/src/main/java/com/careup/ordering/ordering/payment/entity/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.careup.ordering.payment.entity;
+package com.careup.ordering.ordering.payment.entity;
 
 public enum PaymentStatus {
     PENDING, COMPLETED, FAILED, CANCELLED
