+++ ordering/src/main/java/com/careup/ordering/ordering/product/entity/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.product.entity;
+
+public enum ProductStatus {
+    ACTIVE, INACTIVE
+}
