--- ordering/src/main/java/com/careup/ordering/product/entity/ProductStatus.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.product.entity;
-
-public enum ProductStatus {
-    ACTIVE, INACTIVE
-}
